classifier-worker/requirements.txt
arq
redis
sqlalchemy
asyncpg
httpx
openai
python-dotenv
